More internal restructuring
Move more logic top-level, move prints to top-level
