+++ cmd/sbctl/enroll-keys.go
@@ -1,7 +1,11 @@
 package main
 
 import (
+	"errors"
+	"fmt"
+
 	"github.com/foxboron/sbctl"
+	"github.com/foxboron/sbctl/logging"
 	"github.com/spf13/cobra"
 )
 
@@ -9,7 +13,27 @@ var enrollKeysCmd = &cobra.Command{
 	Use:   "enroll-keys",
 	Short: "Enroll the current keys to EFI",
 	RunE: func(cmd *cobra.Command, args []string) error {
-		return sbctl.SyncKeys()
+		var isImmutable bool
+		for _, file := range sbctl.EfivarFSFiles {
+			err := sbctl.IsImmutable(file)
+			if errors.Is(err, sbctl.ErrImmutable) {
+				isImmutable = true
+				logging.Warn("File is immutable: %s", file)
+			} else if errors.Is(err, sbctl.ErrNotImmutable) {
+				continue
+			} else if err != nil {
+				return fmt.Errorf("couldn't read file: %s", file)
+			}
+		}
+		if isImmutable {
+			return sbctl.ErrImmutable
+		}
+		synced := sbctl.SBKeySync(sbctl.KeysPath)
+		if !synced {
+			return errors.New("couldn't sync keys")
+		}
+		logging.Ok("Synced keys!")
+		return nil
 	},
 }
