+++ cmd/sbctl/verify.go
@@ -10,7 +10,6 @@ var verifyCmd = &cobra.Command{
 	Short: "Find and check if files in the ESP are signed or not",
 	RunE: func(cmd *cobra.Command, args []string) error {
 		if err := sbctl.VerifyESP(); err != nil {
-			// Really need to sort out the low level error handling
 			return err
 		}
 		return nil
